line-bot/requirements.txt
fastapi==0.111.0
gunicorn==22.0.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
line-bot-sdk==3.11.0
anthropic==0.28.0
python-dotenv==1.0.1
